rename kicdocker to docker
