+++ pkg/minikube/driver/driver_windows.go
@@ -32,8 +32,10 @@ var supportedDrivers = []string{
 	VMwareFusion,
 	HyperV,
 	VMware,
-	KicDocker,
+	Docker,
 }
+
+// TODO: medyagh add same check for kic docker
 
 func VBoxManagePath() string {
 	cmd := "VBoxManage"
