+++ pkg/minikube/registry/drvs/init.go
@@ -18,9 +18,9 @@ package drvs
 
 import (
 	// Register all of the drvs we know of
+	_ "k8s.io/minikube/pkg/minikube/registry/drvs/docker"
 	_ "k8s.io/minikube/pkg/minikube/registry/drvs/hyperkit"
 	_ "k8s.io/minikube/pkg/minikube/registry/drvs/hyperv"
-	_ "k8s.io/minikube/pkg/minikube/registry/drvs/kicdocker"
 	_ "k8s.io/minikube/pkg/minikube/registry/drvs/kvm2"
 	_ "k8s.io/minikube/pkg/minikube/registry/drvs/none"
 	_ "k8s.io/minikube/pkg/minikube/registry/drvs/parallels"
